% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nafalse.R
\name{nafalse}
\alias{nafalse}
\title{NA False}
\usage{
nafalse(x)
}
\arguments{
\item{x}{a logical vector/array etc.}
}
\value{
the result of x & !is.na(x)
}
\description{
Changes any NAs in a logical object to be false
}
\examples{
try this
}
